Stabilize planning decisions by unifying availability and centralizing domain rules
The production-planning flow had duplicated stock/reception logic and scattered business predicates, which made feasibility, matching, and scheduling drift over time. This commit introduces a shared availability kernel, a centralized domain-rules module, and dedicated diagnostics helpers so calculation paths stay consistent while reporting remains swappable.\n\nConstraint: Repository is in an active multi-area refactor with many unrelated pending changes\nConstraint: Preserve current API/reporting behavior while reducing logic duplication\nRejected: Big-bang rewrite of scheduler/matching/reporting pipeline | too risky for current dirty branch state\nRejected: Keep module-local helper duplication | keeps divergence risk across feasibility/matching/scheduling\nConfidence: high\nScope-risk: moderate\nReversibility: clean\nDirective: Extend business predicates via domain_rules.py first; avoid reintroducing local status/type checks in algorithm modules\nTested: pytest (apps/ordo-core, 201 passed); compile check via python -m compileall production_planning\nNot-tested: Ruff/lint gate (ruff unavailable in current environment)

+++ apps/ordo-core/production_planning/feasibility/feasibility_simulation.py
@@ -41,7 +41,7 @@ class SimulationContext:
         self.stock_state = build_material_stock_state(loader)
 
         # Reception index by day (read-only from real data)
-        self.receptions_by_day: dict[date, list[tuple[str, int]]] = build_receptions_by_day(loader)
+        self.receptions_by_day: dict[date, list[tuple[str, float]]] = build_receptions_by_day(loader)
 
         # Virtual schedule: {date: {poste: hours_used}}
         self.schedule_state: dict[date, dict[str, float]] = defaultdict(lambda: defaultdict(float))
